services/api/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
beautifulsoup4
selectolax
playwright
selenium
python-dotenv
pydantic

sqlalchemy
alembic
psycopg2-binary
redis
rq
celery
apscheduler
passlib[bcrypt]
bcrypt<4

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp

python-jose[cryptography]
pydantic-settings
email-validator
python-multipart
